refactor(platform): replace the missing-impl marker with a registration policy
Stamping a symbol on the registration and deleting it from the .make build
was two mutations papering over one fact. Capture the fact instead:
RequiresImplementationPolicy (the AdoptPolicy/ProviderModePolicy idiom) is
provided at Platform's bare-tag yield site and captured once onto the
registration as RequiresImplementation. Plan.make dies on
RequiresImplementation && Props === undefined — the .make Layer's existing
Props repair makes a satisfied forward reference inert, nothing to erase.

Scoping to platform tags also keeps plain no-props reference yields
(yield* Queue("MyQueue")) planning as noops.
