+++ packages/alchemy/src/Plan.ts
@@ -47,7 +47,7 @@ import {
 } from "./ProviderMode.ts";
 import {
   isResource,
-  MissingImplementationMarker,
+  missingImplementation,
   type ResourceBinding,
   type ResourceLike,
 } from "./Resource.ts";
@@ -300,16 +300,20 @@ export const make = <A>(
     const resources = Object.values(stack.resources);
     const actions = Object.values(stack.actions ?? {});
 
-    // A bare platform tag yielded as a forward reference is stamped with
-    // this marker at registration; its `.make(props, impl)` Layer erases it
-    // when the build repairs the registration's `Props`. A marker that
-    // survived the whole program means the tag was yielded but its Layer was
+    // A bare platform tag yields a forward reference registered with
+    // `undefined` props (and `RequiresImplementation`); its `.make(props,
+    // impl)` Layer repairs the props when it builds (in either order — the
+    // #874 circular env-tag pattern). Props still `undefined` after the
+    // whole program evaluated means the tag was yielded but its Layer was
     // never provided — fail fast naming the class and its Layer instead of
-    // letting a provider read `undefined` props (#1054).
+    // letting a provider read `undefined` props (#1054). Plain resources
+    // may legitimately be yielded without props (a reference to
+    // already-deployed state), so the check is scoped to platform tags.
     for (const resource of resources) {
-      const missingImpl = (resource as any)[MissingImplementationMarker];
-      if (missingImpl !== undefined) {
-        yield* Effect.die(missingImpl);
+      if (resource.RequiresImplementation && resource.Props === undefined) {
+        yield* Effect.die(
+          missingImplementation(resource.Type, resource.LogicalId),
+        );
       }
     }
 
@@ -1842,6 +1846,7 @@ export const make = <A>(
                     Provider: Provider(resourceType),
                     RemovalPolicy: oldState.removalPolicy,
                     Adopt: undefined,
+                    RequiresImplementation: undefined,
                     Mode: oldState.providerMode,
                     FormerFqns: undefined,
                     RuntimeContext: undefined!,
