+++ packages/alchemy/src/Platform.ts
@@ -4,7 +4,6 @@ import type { NodeServices } from "@effect/platform-node/NodeServices";
 import * as ConfigError from "effect/Config";
 import * as ConfigProvider from "effect/ConfigProvider";
 import * as Context from "effect/Context";
-import * as Data from "effect/Data";
 import * as Effect from "effect/Effect";
 import * as Effectable from "effect/Effectable";
 import * as Layer from "effect/Layer";
@@ -23,7 +22,7 @@ import * as Output from "./Output.ts";
 import { ALCHEMY_PHASE } from "./Phase.ts";
 import type { Provider, ProviderCollectionLike } from "./Provider.ts";
 import {
-  MissingImplementationMarker,
+  RequiresImplementationPolicy,
   Resource,
   type ResourceLike,
 } from "./Resource.ts";
@@ -47,54 +46,6 @@ export interface PlatformProps {
    */
   isExternal?: boolean;
 }
-
-/**
- * A tagged platform resource declared with neither props nor an inline
- * implementation was `yield*`ed while its `.make(props, impl)` Layer was
- * absent from the context.
- *
- * Such a tag carries no configuration on its own — props AND impl both live
- * on the Layer — so without it the underlying resource would be registered
- * with `undefined` props and the failure would only surface later, deep
- * inside whichever provider first reads a prop (e.g. `TypeError: undefined
- * is not an object (evaluating 'news.name')` in the Cloudflare Worker
- * pre-create).
- */
-export class MissingImplementationError extends Data.TaggedError(
-  "MissingImplementationError",
-)<{
-  message: string;
-  /** Resource type of the platform, e.g. `Cloudflare.Worker`. */
-  type: string;
-  /** Logical id of the tagged resource (its class name by convention). */
-  id: string;
-}> {}
-
-const missingImplementation = (type: string, id: string) =>
-  new MissingImplementationError({
-    type,
-    id,
-    message: [
-      `${type}<${id}> was yielded without its implementation.`,
-      "",
-      `\`${id}\` is declared as a bare tag — no props, no inline implementation — so both come from its \`.make(...)\` Layer:`,
-      "",
-      `  export class ${id} extends ${type}<${id}>()("${id}") {}`,
-      `  export const ${id}Live = ${id}.make({ /* props */ }, Effect.gen(function* () { /* ... */ }));`,
-      "",
-      `That Layer is not in scope where \`${id}\` is yielded. Provide it to the Stack's program:`,
-      "",
-      "  Alchemy.Stack(",
-      `    "my-stack",`,
-      "    { providers, state },",
-      "    Effect.gen(function* () {",
-      `      const instance = yield* ${id};`,
-      `    }).pipe(Effect.provide([${id}Live])),`,
-      "  )",
-      "",
-      `If \`${id}\` is not Effect-native, declare it with props instead: \`()("${id}", { /* props */ })\`.`,
-    ].join("\n"),
-  });
 
 /**
  * Provide the platform class's layer (`cls.make(props, impl)`) with a
@@ -446,39 +397,25 @@ export const Platform = <
                 // — same isExternal marking; without props, this is a bare
                 // tag whose props/impl arrive later via `.make`.
                 onNone: () =>
-                  props === undefined && !globalThis.__ALCHEMY_RUNTIME__
-                    ? Effect.map(
-                        resource(id, applyTransformProps(id, props)),
-                        (instance: any) => {
-                          // A bare tag yielded here is a FORWARD REFERENCE:
-                          // its `.make(props, impl)` Layer may register the
-                          // real props before or after this yield (e.g. a
-                          // worker tag bound in another worker's `env` — the
-                          // #874 circular-binding pattern — resolves during
-                          // that worker's async-binding pass, outside the
-                          // Layer's own context and possibly before the
-                          // Layer builds). So this site cannot decide whether
-                          // the Layer is missing. Stamp the registration
-                          // instead; the Layer's build erases the marker when
-                          // it repairs `Props`, and `Plan.make` dies on any
-                          // marker that survives — naming the class and its
-                          // Layer instead of letting a provider read
-                          // `undefined` props. Guarded to plan/deploy so the
-                          // runtime bridges keep their current behaviour (and
-                          // the branch is folded out of deployed bundles).
-                          if (instance.Props === undefined) {
-                            instance[MissingImplementationMarker] =
-                              missingImplementation(type, id);
-                          }
-                          return instance;
-                        },
+                  // Without props this is a bare-tag FORWARD REFERENCE: its
+                  // `.make(props, impl)` Layer may build before or after
+                  // this yield (e.g. a worker tag bound in another worker's
+                  // `env` — the #874 circular-binding pattern — resolves
+                  // during that worker's async-binding pass, outside the
+                  // Layer's own context). Register with `undefined` props;
+                  // the Layer's build repairs them, and `Plan.make` fails
+                  // fast on any `RequiresImplementation` registration whose
+                  // props are still `undefined` after the whole program
+                  // evaluated — the tag was yielded but its Layer was never
+                  // provided (#1054).
+                  props === undefined
+                    ? resource(id, applyTransformProps(id, props)).pipe(
+                        Effect.provideService(
+                          RequiresImplementationPolicy,
+                          true,
+                        ),
                       )
-                    : resource(
-                        id,
-                        props === undefined
-                          ? applyTransformProps(id, props)
-                          : externalProps(),
-                      ),
+                    : resource(id, externalProps()),
                 onSome: Effect.succeed,
               }),
             )
@@ -717,10 +654,6 @@ export const Platform = <
                   ? yield* runtimeContext.exports
                   : undefined,
               };
-              // This build supplied the tag's props/impl — a bare-tag yield
-              // that ran BEFORE this Layer built (a forward reference, e.g.
-              // the #874 circular env-tag pattern) is now satisfied.
-              delete (instance as any)[MissingImplementationMarker];
 
               return Object.assign(instance, {
                 RuntimeContext: runtimeContext,
